Refactor - move web app src code to app folder.
