+++ src/web/.env_example
@@ -1,4 +1,4 @@
-API_URL=http://127.0.0.1:8000
+BASE_WEB_API_URL=http://127.0.0.1:8001
 BLOB_STORAGE_BBLOB_STORAGE_BASE_URL=https://footytrackerstoragedev.blob.core.windows.net
 BLOB_STORAGE_CON_STR=
 BLOB_PROFILE_IMAGE_CONTAINER=profileimages
