+++ src/web/app/infrastructure/cookie_auth.py
