+++ src/web/app/static/img/sword.png
