+++ src/web/app/viewmodels/account/edit_viewmodel.py
@@ -1,11 +1,10 @@
-from typing import Optional, List
+from typing import Optional
 
 from starlette.requests import Request
 
-from models.user import UserRead, UserUpdate
-from models.validation_error import ValidationError
-from services import user_service
-from viewmodels.shared.viewmodel import ViewModelBase
+from app.models.user import UserRead, UserUpdate
+from app.services import user_service
+from app.viewmodels.shared.viewmodel import ViewModelBase
 
 
 class AccountEditViewModel(ViewModelBase):
