+++ src/web/app/viewmodels/account/register_viewmodel.py
@@ -1,11 +1,10 @@
 from typing import Optional
-import ast
 
 from starlette.requests import Request
 
-from models.validation_error import ValidationError
-from services import user_service
-from viewmodels.shared.viewmodel import ViewModelBase
+from app.models.validation_error import ValidationError
+from app.services import user_service
+from app.viewmodels.shared.viewmodel import ViewModelBase
 
 
 class RegisterViewModel(ViewModelBase):
